+++ CHANGELOG.md
@@ -52,6 +52,7 @@
     -   Records actions, like `os.recordData()`, `os.getData()`, `os.recordFile()`, etc. can be passed an endpoint which specifies which backend should be used for the request.
     -   If no endpoint is specified, then a default is used.
     -   `os.getRecordsEndpoint()` returns a promise that resolves to the endpoint that is used by default.
+-   Added the `os.showAccountInfo()` function to show the "Account Information" dialog for users who are logged in.
 
 ### :bug: Bug Fixes